version 4 tp3

+++ tp3v3/TP3/api.ts
@@ -10,7 +10,7 @@ const pool = mysql.createPool({    //si van a haber procesos paralelos hay q usa
     connectionLimit: 10,   //max de hilos  
     host: "localhost",
     user: "root",
-    password: "password",
+    password: "",
     database: "cine",
 });
 
@@ -25,7 +25,7 @@ const stringAArray = (string: String): Array<String> =>{
     return arrayNuevo;    
 };
 const arrayAString = (array: Array<String>): String =>{
-    let nuevoString = `["`+array[0]+`"`;
+    let nuevoString = `["${array[0]}"`;
     
     for(let i: number = 1; i < array.length; i++){
         nuevoString += `, "${array[i]}"`;
@@ -37,9 +37,7 @@ const arrayAString = (array: Array<String>): String =>{
 
 if(cluster.isWorker){
     
-    process.on('message', (bodyReserva) => {
-
-        //pool.query --> pool.getConnection conn.query conn.release()
+    process.on('message', (bodyReserva, idFuncion) => {
         
         pool.getConnection( (error1, conn) => {
             if (error1) {
@@ -47,28 +45,21 @@ if(cluster.isWorker){
             }
             conn.beginTransaction( (error2) => {
                 if (error2) {throw error2}
-                //Supongo que aparte me dan el id de la funcion
-                //Disponibilidad de butacas listo
-                //Generar reserva oki
-                //Si se llena la funcion, marcar como no diponible
 
                 let idUsuario:number = bodyReserva.user_id;
-                let idFuncion:number = bodyReserva.funcion_id;
-                
-                let arrayButacas:Array <String> = stringAArray(bodyReserva.butacas);                              
-                console.log(arrayButacas); //esta ya es array
+                let arrayButacas:Array <String> = stringAArray(bodyReserva.butacas);
                 
                 if (arrayButacas.length >= 6){
+                    
                     process.kill(process.pid);    
                 }
-
                  
                 conn.commit(() => {
                     conn.release(); 
                     process.kill(process.pid);
                 });
                 
-                /*conn.query(`select vigente from funciones where id=${idFuncion}`, (error,results) =>{
+                conn.query(`select vigente from funciones where id=${idFuncion}`, (error,results) =>{
                     if (results) {
                         conn.query(`select count(id) from reservas where usuario=${idUsuario} and funcion=${idFuncion}`, (error,results) =>{
                             if(results==0) {
@@ -76,40 +67,45 @@ if(cluster.isWorker){
                                     if(error){
                                         throw error;
                                     }
-                                    let arrayButacasDisponibles:Array<String> = stringAArray(results[0].butacas_disponibles);
-                                    let butacasDisponiblesParaReservar:boolean=true;
-                                    arrayButacas.forEach(butaca =>{
-                                        if(!arrayButacasDispone.includes(butaca)){
-                                            butacasDisponiblesParaReservar=false;
+                                    let arrayButacasDisponibles: Array<String> = stringAArray(results[0].butacas_disponibles);
+                                    let butacasDisponiblesParaReservar: boolean = true;
+                                    arrayButacas.forEach(butaca => {
+                                        if(!arrayButacasDisponibles.includes(butaca)){
+                                            butacasDisponiblesParaReservar = false;
                                         }
-                                    })
+                                    });
                                     if(butacasDisponiblesParaReservar){
-                                        conn.query(`insert into `)
-                                        ahora tendria que reservar entonces
-                                        y luego tendria que chequear si tengo que marcar la funcion como no disponible
+                                        conn.query(`insert into reservas value (null, ${idUsuario}, ${idFuncion}, ${arrayButacas})`, (error) => {
+                                            if (error) throw error;
+                                            else {
+                                                let butacasActualizadas: Array<String> = arrayButacasDisponibles;
+                                                
+                                                arrayButacas.forEach(butaca => {
+                                                    butacasActualizadas.splice(butacasActualizadas.indexOf(butaca),1);
+                                                });
+
+                                                if (butacasActualizadas.length > 0){
+                                                    conn.query(`update funciones set butacas_dispoibles = ${arrayAString(butacasActualizadas)}`, (error) => {
+                                                        if (error) throw error;
+                                                    });
+                                                } else {
+                                                    conn.query(`update funciones set butacas_dispoibles = '[]', vigente = 0`, (error) => {
+                                                        if (error) throw error;
+                                                    });
+                                                }
+                                            }
+                                        });
                                     } 
+                                });
                             }
-                        })
+                        });
                     }
-                })*/
-                /*conn.query("", (error3, results) =>{
-                    if (error3) {
-                        return conn.rollback(() => {throw error3})
-                    }
-                    conn.commit((error4) =>{
-                        if (error4){
-                            return conn.rollback(()=>{
-                                throw error4
-                            })
-                        }
-                    })
-                    res.send("Insertado con exito");
-                })*/
-            });            
+                });
+            });
         });
     });
-} else {
 
+} else {
     const port:number = 3305; 
     const cluster = require('cluster');
     const bodyParser = require('body-parser');
@@ -129,15 +125,13 @@ if(cluster.isWorker){
         });
     })
     
-    
-    app.post('/reservar', (req,res) =>{
+    app.post('/{id_funcion}/reservar', (req,res) =>{
         const worker = cluster.fork();
         
-        worker.send(req.body);
+        worker.send(req.body, req.query.id_funcion);
         worker.on('message', (result) => {
             res.status(200).send(result);
         });
-        
     });    
     
     
@@ -145,33 +139,37 @@ if(cluster.isWorker){
         console.log(`App listening at http://localhost:${port}`)
     });
 
-    
-    app.post('/cancelar_reserva', (req,res) =>{
-        let user_id:number=req.body.user_id; 
-        let id_reserva:number=req.body.id_reserva;
-        pool.query(`select butacas_reservadas from reservas inner join funciones on reservas.funcion=funciones.id where reservas.id=${id_reserva} && (date_add(now(),interval 1 hour))<funciones.fecha`, async (error,results) =>{
-        if(error) throw error;
-        else if (results.length > 0){
-            await pool.query(`SELECT butacas_disponibles FROM funciones inner join reservas on funciones.id=reservas.funcion WHERE reservas.id=${id_reserva}`, async (error2,results2) =>{
-                if(error2) throw error;
-                else{  
-                    console.log("MEGA UWU");
-                    let butacasReservadas: Array<String> = stringAArray(results[0].butacas_reservadas);
-                    let butacasDisponibles: Array<String> = stringAArray(results2[0].butacas_disponibles);
-                    console.log(butacasReservadas);
-                    console.log(butacasDisponibles);
-                    butacasDisponibles=butacasDisponibles.concat(butacasReservadas);
-                    let butacasDisaponiblesString : String = arrayAString(butacasDisponibles);
-                    console.log(butacasDisaponiblesString);
-                    await pool.query(`update funciones join reservas on funciones.id=reservas.funcion set butacas_disponibles='${butacasDisaponiblesString}' where reservas.id=${id_reserva}`);
-                    await pool.query(`delete from reservas where id=${id_reserva}`);
-                }
-            });
-            res.send("Eliminado correctamente");
-        }
-        else{res.send("El lapso para cancelar la reserva termino o la reserva no exite.");}
+    app.post('/{id_funcion}/cancelar_reserva', (req,res) =>{
+        let id_funcion:number = req.query.id_reserva;
+        let user_id:number = req.body.user_id;
+
+        pool.query(`select id from reservas where usuario = ${user_id} && funcion = ${id_funcion}`,async (error,results) =>{
+            if(error) throw error;
+            else if (results.length > 0){
+                let id_reserva: number = results[0].id;
+                pool.query(`select butacas_reservadas from reservas inner join funciones on reservas.funcion=funciones.id where reservas.id=${id_reserva} && (date_add(now(),interval 1 hour))<funciones.fecha`, async (error2,results) =>{
+                    if(error2) throw error;
+                    else if (results.length > 0){
+                        await pool.query(`SELECT butacas_disponibles FROM funciones inner join reservas on funciones.id=reservas.funcion WHERE reservas.id=${id_reserva}`, async (error3,results2) =>{
+                            if(error3) throw error;
+                            else{  
+                                console.log("MEGA UWU");
+                                let butacasReservadas: Array<String> = stringAArray(results[0].butacas_reservadas);
+                                let butacasDisponibles: Array<String> = stringAArray(results2[0].butacas_disponibles);
+                                console.log(butacasReservadas);
+                                console.log(butacasDisponibles);
+                                butacasDisponibles=butacasDisponibles.concat(butacasReservadas);
+                                let butacasDisaponiblesString : String = arrayAString(butacasDisponibles);
+                                console.log(butacasDisaponiblesString);
+                                await pool.query(`update funciones join reservas on funciones.id=reservas.funcion set butacas_disponibles='${butacasDisaponiblesString}' where reservas.id=${id_reserva}`);
+                                await pool.query(`delete from reservas where id=${id_reserva}`);
+                            }
+                        });
+                        res.send("Eliminado correctamente");
+                    }
+                    else{res.send("El lapso para cancelar la reserva termino o la reserva no exite.");}
+                });
+            }
         });
-
     });
-
 }
